requirements.txt
python-telegram-bot==13.7
telegram==0.0.1
python-dotenv==0.19.0
